MAINT moving methods around to avoid circular imports.
The API now doesn't make sense anymore. @choldgraf and @NelleV should now sit
down and start discussing what needs to be where.
