+++ watchtower/handlers_.py
@@ -246,116 +246,3 @@ class ProjectActivity(object):
         s = 'Project: {} / {} | N Issues: {} | N Commits: {}'.format(
             self.user, self.project, n_issues, n_commits)
         return s
-
-
-class ProjectCommits(object):
-    """Represent commit activity for a github project.
-
-    This stores information about commit activity.
-
-    Parameters
-    ----------
-    user : string
-        The user to query
-    project : string
-        The repository to query.
-    raw : json return by github api
-        The raw data for this project, as returned by the github api.
-
-    Attributes
-    ----------
-    commits : DataFrame
-        A collection of commit activity for this project.
-    """
-    def __init__(self, user, project, raw):
-        self.user = user
-        self.project = project
-        self.raw = raw
-
-        # Extract commits information that is useful
-        commits = dict(email=[], name=[], message=[], sha=[],
-                       date=[])
-
-        for commit, sha, date in raw[['commit', 'sha', 'date']].values:
-            commits['email'].append(commit['author']['email'])
-            commits['name'].append(commit['author']['name'])
-            commits['message'].append(commit['message'])
-            commits['sha'].append(sha)
-            commits['date'].append(date)
-        commits = pd.DataFrame(commits)
-        commits = commits.set_index('date')
-        commits.index = pd.to_datetime(commits.index)
-        commits['project'] = project
-        commits['user'] = user
-        self.commits = commits
-
-    def __repr__(self):
-        s = 'Project: {} / {} | N Events: {}'.format(
-            self.user, self.project, len(self.commits))
-        return s
-
-
-class ProjectIssues(object):
-    """Represent issue for a github project.
-
-    This stores information about issue activity over time.
-
-    Parameters
-    ----------
-    user : string
-        The user to query
-    project : string
-        The repository to query.
-    raw : json return by github api
-        The raw data for this project, as returned by the github api.
-
-    Attributes
-    ----------
-    issues : DataFrame
-        A collection of issues activity for this project.
-    label_names : list of strings
-        A list of strings representing each label in the project
-    """
-    def __init__(self, user, project, raw):
-        self.user = user
-        self.project = project
-        raw = raw.set_index('date')
-        raw.index = pd.to_datetime(raw.index)
-
-        self.raw = raw
-
-        # Extract issue information that is useful
-        # Building a dictionary in case we need to dig
-        # deeper in future versions of watchtower
-        issues = dict(
-            title=[], labels=[], state=[],
-            updated_at=[], html_url=[], id=[],
-            user=[], pull_request=[])
-        if len(raw) == 0:
-            self.issues = None
-            return
-        use_keys = list([key for key in issues.keys()
-                         if key in raw.columns])
-        insert_keys = list([key for key in issues.keys()
-                            if key not in raw.columns])
-        issues = raw[use_keys]
-
-        # Label names
-        label_names = [list(label['name']
-                            for label in labels)
-                       for labels in issues['labels'].values]
-        issues['label_names'] = label_names
-        for key in insert_keys:
-            issues[key] = np.nan
-
-        # Has a PR
-        has_pr = -1 * issues['pull_request'].isnull()
-        issues['has_pr'] = has_pr
-        self.issues = issues
-
-    def __repr__(self):
-        s = 'Issues: {} / {} | N Open: {} | N Closed: {}'.format(
-            self.user, self.project,
-            sum(self.issues['state'] == 'open'),
-            sum(self.issues['state'] == 'closed'))
-        return s
