+++ watchtower/issues_.py
@@ -2,10 +2,10 @@ import os
 from os.path import join
 
 import pandas as pd
+import numpy as np
 
 from . import _github_api
 from ._config import get_data_home, get_API_token, DATETIME_FORMAT
-from .handlers_ import ProjectIssues
 
 
 def update_issues(user, project, auth=None, state="all", since=None,
@@ -105,3 +105,69 @@ def load_issues(user, project, data_home=None,
         return issues
     else:
         return ProjectIssues(user, project, issues)
+
+
+class ProjectIssues(object):
+    """Represent issue for a github project.
+
+    This stores information about issue activity over time.
+
+    Parameters
+    ----------
+    user : string
+        The user to query
+    project : string
+        The repository to query.
+    raw : json return by github api
+        The raw data for this project, as returned by the github api.
+
+    Attributes
+    ----------
+    issues : DataFrame
+        A collection of issues activity for this project.
+    label_names : list of strings
+        A list of strings representing each label in the project
+    """
+    def __init__(self, user, project, raw):
+        self.user = user
+        self.project = project
+        raw = raw.set_index('date')
+        raw.index = pd.to_datetime(raw.index)
+
+        self.raw = raw
+
+        # Extract issue information that is useful
+        # Building a dictionary in case we need to dig
+        # deeper in future versions of watchtower
+        issues = dict(
+            title=[], labels=[], state=[],
+            updated_at=[], html_url=[], id=[],
+            user=[], pull_request=[])
+        if len(raw) == 0:
+            self.issues = None
+            return
+        use_keys = list([key for key in issues.keys()
+                         if key in raw.columns])
+        insert_keys = list([key for key in issues.keys()
+                            if key not in raw.columns])
+        issues = raw[use_keys]
+
+        # Label names
+        label_names = [list(label['name']
+                            for label in labels)
+                       for labels in issues['labels'].values]
+        issues['label_names'] = label_names
+        for key in insert_keys:
+            issues[key] = np.nan
+
+        # Has a PR
+        has_pr = -1 * issues['pull_request'].isnull()
+        issues['has_pr'] = has_pr
+        self.issues = issues
+
+    def __repr__(self):
+        s = 'Issues: {} / {} | N Open: {} | N Closed: {}'.format(
+            self.user, self.project,
+            sum(self.issues['state'] == 'open'),
+            sum(self.issues['state'] == 'closed'))
+        return s
